requirements.txt
Flask==3.0.0
gunicorn==21.2.0
openai>=1.0.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0